+++ Source/Game/MouseManager.cpp
@@ -15,7 +15,7 @@ void MouseManager::Initialize(CommonUtilities::GrowingArray<std::string> &aFileP
 	myInputManager = aInputManager;
 
 	mySpriteInteractive.Init(6); // There are six different cursors
-	for (int i = 0; i < aFilePath.Size(); i++)
+	for (unsigned short i = 0; i < aFilePath.Size(); i++)
 	{
 		mySprite = new DX2D::CSprite(aFilePath[i].c_str());
 		mySpriteInteractive.Add(mySprite);
@@ -39,9 +39,9 @@ void MouseManager::Initialize(CommonUtilities::GrowingArray<std::string> &aFileP
 void MouseManager::Update(float)
 {
 	myInputManager->SetAbsoluteMousePos(
-		(ResolutionManager::GetInstance()->GetRenderAreaDimension().x + ResolutionManager::GetInstance()->GetRenderAreaPosition().x) * 0.5f,  
-		(ResolutionManager::GetInstance()->GetRenderAreaDimension().y + ResolutionManager::GetInstance()->GetRenderAreaPosition().y) * 0.5f);
-		
+		static_cast<int>((ResolutionManager::GetInstance()->GetRenderAreaDimension().x + ResolutionManager::GetInstance()->GetRenderAreaPosition().x) * 0.5f),  
+		static_cast<int>((ResolutionManager::GetInstance()->GetRenderAreaDimension().y + ResolutionManager::GetInstance()->GetRenderAreaPosition().y) * 0.5f));
+	
 	float aSpeed = 0.0005f;
